// src/app/core/dataservice/lease/lease-agreement.dto.ts
export interface LeaseAgreeementDTO {
    id: number;
    type: LEASETYPE;
    adminId: number;

    bankAccountId: number;
    bankAccount: BankAccountDto;

    status: LEASESTATUS;
    lesseeType: LESSEETYPE;
    lessorType: LESSORTYPE;

    entryDamageReportSubmitted: boolean;
    securityDepositPaid: boolean;

    leaseDurationMonths: number;
    leaseStartDate: string;
    leaseSigningDate: string;
    leaseModificationDate: string;
    leaseModificationRemarks: string;
    leaseEndDate: string;

    plotId: number;
    plot: PlotDTO;

    tenantId: number;
    tenant: UserDTO;

    organizationId?: number;
    organization?: OrganiztionDTO;
    buildingId?: number;
    building?: BuildingDTO;
    unitId?: number;
    unit?: UnitDTO;

    rent: number;
    use: LEASEUSES;

    securityDepositAmount: number;
    paymentDueDay: number;
    applyLatePaymentFee: boolean;

    tenantSubletAuthority: boolean;
    tenantPrematureTermination: boolean;
    ownerPrematureTermination: boolean;

    vacationNoticePeriod: number;
    evictionNoticePeriod: number;
    rentIncreaseNoticePeriod: number;

    penaltyPercentagePerAnnum: number;
    rentIncrementPercentage: number;
    rentIncrementDurationYear: number;

    areaLeased?: number;
    areaUnit?: string;
    ratePerArea?: number;

    leaseSurcharges: LeaseSurchargeDTO[];
    leaseRules: LeaseRuleDTO[];
    paymentAdvises?: PaymentAdviceDto[];
}